Removed silent authentication
Authentication service logs the user out of their Auth0 session when they press the logout button so that they will have to log in again.

If the user closes the tab when they are still logged in, the next time they go to the SPWA, they will be logged out automatically so that they will have to log in again.